refactor(email): migrate callers off the digest shims, delete shims
Move yas.worker.digest_loop, yas.web.routes.digest_preview, and the
remaining tests off the yas.alerts.digest.{builder,filters} re-export
shims onto yas.email directly. Two production callers now use
render_digest_payload() (which returns RenderedEmail) and unpack
.body_plain/.body_html locally -- the worker keeps the explicit
'Daily digest — {kid} — {date}' subject format so the migration is
behavior-preserving (rendered.subject would be just the top_line).

Delete the now-unused shim modules:
  src/yas/alerts/digest/builder.py
  src/yas/alerts/digest/filters.py

src/yas/alerts/digest/llm_summary.py stays where it is -- it is not a
shim, it is the LLM top-line generator. A future cleanup may move it
under yas.email but that is out of scope here.

Two existing test files (test_alerts_digest_builder.py,
test_digest_golden.py) keep their historical filenames but gain a small
local render_digest() helper that wraps render_digest_payload back into
the (plain, html) tuple shape their assertions were written against.
