--- src/yas/alerts/digest/filters.py
@@ -1,10 +1,0 @@
-"""Re-export shim — filters moved to yas.email.filters in Task 4.
-
-Kept so existing callers (and historical imports) continue to resolve. New code
-should import from yas.email.filters directly.
-"""
-from __future__ import annotations
-
-from yas.email.filters import fmt, price, rel_date
-
-__all__ = ["fmt", "price", "rel_date"]
