+++ tests/unit/test_alerts_digest_llm_summary.py
@@ -9,8 +9,8 @@ from typing import Any
 
 import pytest
 
-from yas.alerts.digest.builder import DigestPayload
 from yas.alerts.digest.llm_summary import generate_top_line
+from yas.email.payloads import DigestPayload
 
 # ---------------------------------------------------------------------------
 # Test double
